+++ src/lib.rs
@@ -31,6 +31,13 @@ pub mod player_context;
 /// Re-exports of all the most common types.
 pub mod prelude;
 
+/// A macro that transforms `async` functions (and closures) into plain functions, whose return
+/// type is a boxed [`Future`].
+///
+/// [`Future`]: std::future::Future
+#[cfg(feature = "macros")]
+pub use macros_dep::hook;
+
 #[cfg(feature = "python")]
 use pyo3::{prelude::*, types::PyDict, wrap_pymodule};