+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Remove typemap_rev and parking_lot dependencies.
 - Remove user-data feature.
 - Move user-data to a standard rust implementation inspired by 0.13 serenity.
+- Merged `serenity-native` and `serenity-rustls` into a single `serenity` feature.
 
 ## 0.10.0